Split the interface from the matchers and add Any.of() + All.of()
 * We now have an interface providing Any and All
 * All the actual matchers are elsewhere
 * This means we can import them into each other without loops
