+++ src/h_matchers/__init__.py
@@ -1,3 +1,3 @@
 """Exposes the user facing matchers for importing."""
 
-from h_matchers.matcher.any import Any
+from h_matchers.interface import All, Any
